third_party/blink/web_tests/fast/dom/HTMLImageElement/image-sizes-2x-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS document.getElementById("simple").clientWidth is window.innerWidth
PASS currentSrcFileNameNoParams("simple") is "image-set-2x.png"
PASS document.getElementById("small_sizes").clientWidth is 300
PASS currentSrcFileNameNoParams("small_sizes") is "image-set-2x.png"
PASS document.getElementById("sizes_x_descriptors").clientWidth is 400
PASS currentSrcFileNameNoParams("sizes_x_descriptors") is "image-set-2x.png"
PASS document.getElementById("sizes_src").clientWidth is 400
PASS currentSrcFileNameNoParams("sizes_src") is "image-set-1x.png"
PASS document.getElementById("default_sizes").clientWidth is window.innerWidth
PASS currentSrcFileNameNoParams("default_sizes") is "image-set-4x.png"
PASS document.getElementById("invalid_length").clientWidth is window.innerWidth
PASS currentSrcFileNameNoParams("invalid_length") is "image-set-2x.png"
PASS document.getElementById("src_and_w").clientWidth is within 0.015625 of 700
PASS currentSrcFileNameNoParams("src_and_w") is "image-set-4x.png"
PASS document.getElementById("x_and_w").clientWidth is within 0.015625 of 700
PASS currentSrcFileNameNoParams("x_and_w") is "image-set-4x.png"
PASS document.getElementById("x_and_w2").clientWidth is 400
PASS currentSrcFileNameNoParams("x_and_w2") is "image-set-2x.png"
PASS document.getElementById("x_and_w3").clientWidth is 400
PASS currentSrcFileNameNoParams("x_and_w3") is "image-set-2x.png"
PASS document.getElementById("half_size").clientWidth is (window.innerWidth/2)
PASS currentSrcFileNameNoParams("half_size") is "image-set-2x.png"
PASS document.getElementById("infitisimal_sizes").clientWidth is (window.innerWidth/100)
PASS currentSrcFileNameNoParams("infitisimal_sizes") is "image-set-2x.png"
           
